HWS: Parallelize STE parsing
We now parse STEs on multiple cores, in 64k chunks. The resulting output
is no longer deterministic by default, but we added the --ordered_output
command line switch to force it to be, for puposes of diff testing.

Measurements taken on a 4M STE dump, excluding FW dump time (so, with
--skip_dump):

			Wall time		Max RSS
Before			3:24			12.0G
This patch		0:57			 8.3G
This patch, ordered	0:57			 8.5G

Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>
Signed-off-by: Vlad Dogaru <vdogaru@nvidia.com>